Codigo/requirements.txt
python-dotenv
pandas
requests
gitpython